Remove /app/(site) and use /app only for all pages routes
