+++ app/dashboard/admin/[slug]/page.tsx
